Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -7,8 +7,8 @@ int main() {
     scanf("%d", &c);
 
     int sum = (a+b+c);
-    float avg;
-    avg= (sum/3.0);
+    double avg;
+    avg= (sum/3);
     printf("Average: %.2f",avg);
     return 0;
 }
